refactor(subway): replace legacy StationDao and StationService
- remove all unused code
