--- src/main/java/wooteco/subway/dao2/StationDao2.java
@@ -1,92 +1,0 @@
-package wooteco.subway.dao2;
-
-import java.util.List;
-import java.util.Optional;
-import org.springframework.jdbc.core.RowMapper;
-import org.springframework.jdbc.core.namedparam.BeanPropertySqlParameterSource;
-import org.springframework.jdbc.core.namedparam.EmptySqlParameterSource;
-import org.springframework.jdbc.core.namedparam.MapSqlParameterSource;
-import org.springframework.jdbc.core.namedparam.NamedParameterJdbcTemplate;
-import org.springframework.jdbc.core.namedparam.SqlParameterSource;
-import org.springframework.jdbc.support.GeneratedKeyHolder;
-import org.springframework.jdbc.support.KeyHolder;
-import org.springframework.stereotype.Repository;
-import wooteco.subway.domain.Station;
-
-@Repository
-public class StationDao2 {
-
-    private static final RowMapper<Station> stationRowMapper = (resultSet, rowNum) ->
-            new Station(resultSet.getLong("id"),
-                    resultSet.getString("name"));
-
-    private final NamedParameterJdbcTemplate jdbcTemplate;
-
-    public StationDao2(NamedParameterJdbcTemplate jdbcTemplate) {
-        this.jdbcTemplate = jdbcTemplate;
-    }
-
-    public List<Station> findAll() {
-        final String sql = "SELECT * FROM station";
-
-        return jdbcTemplate.query(sql, new EmptySqlParameterSource(), stationRowMapper);
-    }
-
-    public boolean checkById(Long id) {
-        final String sql = "SELECT * FROM station WHERE id = :id";
-        MapSqlParameterSource paramSource = new MapSqlParameterSource();
-        paramSource.addValue("id", id);
-
-        return isPresent(jdbcTemplate.query(sql, paramSource, stationRowMapper));
-    }
-
-    public Optional<Station> findById(Long id) {
-        final String sql = "SELECT * FROM station WHERE id = :id";
-        MapSqlParameterSource paramSource = new MapSqlParameterSource();
-        paramSource.addValue("id", id);
-
-        return jdbcTemplate.query(sql, paramSource, stationRowMapper)
-                .stream()
-                .findFirst();
-    }
-
-    public Optional<Station> findByName(String name) {
-        final String sql = "SELECT * FROM station WHERE name = :name";
-        MapSqlParameterSource paramSource = new MapSqlParameterSource();
-        paramSource.addValue("name", name);
-
-        return jdbcTemplate.query(sql, paramSource, stationRowMapper)
-                .stream()
-                .findFirst();
-    }
-
-    public boolean checkByName(String name) {
-        final String sql = "SELECT * FROM station WHERE name = :name";
-        MapSqlParameterSource paramSource = new MapSqlParameterSource();
-        paramSource.addValue("name", name);
-
-        return isPresent(jdbcTemplate.query(sql, paramSource, stationRowMapper));
-    }
-
-    public Station save(Station station) {
-        final String sql = "INSERT INTO station(name) VALUES (:name)";
-        KeyHolder keyHolder = new GeneratedKeyHolder();
-        SqlParameterSource paramSource = new BeanPropertySqlParameterSource(station);
-
-        jdbcTemplate.update(sql, paramSource, keyHolder);
-        Number generatedId = keyHolder.getKey();
-        return new Station(generatedId.longValue(), station.getName());
-    }
-
-    public void deleteById(Long id) {
-        final String sql = "DELETE FROM station WHERE id = :id";
-        MapSqlParameterSource paramSource = new MapSqlParameterSource();
-        paramSource.addValue("id", id);
-
-        jdbcTemplate.update(sql, paramSource);
-    }
-
-    private boolean isPresent(List<Station> stations) {
-        return !stations.isEmpty();
-    }
-}
